requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.6
datasets>=2.14.0
evaluate>=0.4.0
pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0